+++ grid2op/Opponent/RandomLineOpponent.py
@@ -117,10 +117,9 @@ class RandomLineOpponent(BaseOpponent):
         status = observation.line_status[self._lines_ids]
 
         # If all attackable lines are disconnected
-        if np.all(status == False):
+        if np.all(~status):
             return None  # i choose not to attack in this case
 
         # Pick a line among the connected lines
         attack = self.space_prng.choice(self._attacks[status])
-        print(f"attack: {attack}")
         return attack